Projectile does stuff. Somehow made the player seizures a little better.

+++ MansionMashers/MansionMashers/HeaderFiles/Enemy.h
@@ -26,7 +26,7 @@ typedef struct EnemyParts
 	float AttackRotation;
 	float AttackRotationArm;
 	float AttackRotationArmLower;
-	int BlinkTimer;
+	float BlinkTimer;
 }EnemyParts;
 
 typedef struct EnemyStats
